Split WebSocket hub tests by behavior
